moondream2/api/requirements.txt
flask
flask-restx
transformers
Pillow
torch
